Move abstract base classes into "api/items" folder, so that "api/model" contains only the non-abstract classes
