+++ apps/api-extractor/src/api/model/ApiMethodSignature.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { ApiItemKind } from './ApiItem';
-import { IApiMethodItemOptions, ApiMethodItem } from './ApiMethodItem';
+import { ApiItemKind } from '../items/ApiItem';
+import { IApiMethodItemOptions, ApiMethodItem } from '../items/ApiMethodItem';
 
 /** @public */
 export interface IApiMethodSignatureOptions extends IApiMethodItemOptions {
